refactor(protocol): Rename utils/StreamPartitionID to utils/SPID.
