+++ packages/protocol/src/utils/index.ts
@@ -3,12 +3,12 @@ import OrderingUtil from "./OrderingUtil"
 import StreamMessageValidator from "./StreamMessageValidator"
 import CachingStreamMessageValidator from "./CachingStreamMessageValidator"
 import SigningUtil from "./SigningUtil"
-import StreamPartitionID, { SPIDish } from "./StreamPartitionID"
+import SPID, { SPIDish } from "./SPID"
 import { createTrackerRegistry, getTrackerRegistryFromContract, TrackerRegistry } from "./TrackerRegistry"
 
 export {
     SPIDish,
-    StreamPartitionID,
+    SPID,
     TimestampUtil,
     OrderingUtil,
     StreamMessageValidator,
